+++ lista_encadeada/test_linked_list.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-
 import unittest
 import linked_list
